test/unit/modules/content_verifier/commands/create_content_command.spec.ts
import { CreateContentCommand } from '../../../../../src/app/modules/contentVerifier/commands/create_content_command';

describe('CreateContentCommand', () => {
  let command: CreateContentCommand;

	beforeEach(() => {
		command = new CreateContentCommand();
	});

	describe('constructor', () => {
		it('should have valid name', () => {
			expect(command.name).toEqual('createContent');
		});

		it('should have valid schema', () => {
			expect(command.schema).toMatchSnapshot();
		});
	});

	describe('verify', () => {
		describe('schema validation', () => {
      it.todo('should throw errors for invalid schema');
      it.todo('should be ok for valid schema');
    });
	});

	describe('execute', () => {
    describe('valid cases', () => {
      it.todo('should update the state store');
    });

    describe('invalid cases', () => {
      it.todo('should throw error');
    });
	});
});
